Introduce kernel module
Move the KernelResolver functionality into the kernel module. Anything
is better than having it in the crate root.
