+++ src/lib.rs
@@ -14,20 +14,17 @@
 extern crate test;
 
 use std::fmt::Debug;
-use std::fmt::Formatter;
-use std::fmt::Result as FmtResult;
 use std::io::Error;
-use std::io::ErrorKind;
 use std::io::Result;
 use std::path::Component;
 use std::path::Path;
 use std::path::PathBuf;
-use std::rc::Rc;
 
 mod c_api;
 mod dwarf;
 mod elf;
 mod gsym;
+mod kernel;
 mod ksym;
 mod maps;
 mod mmap;
@@ -39,8 +36,8 @@ mod zip;
 use elf::ElfCache;
 use elf::ElfResolver;
 use gsym::GsymResolver;
+use kernel::KernelResolver;
 use ksym::KSymCache;
-use ksym::KSymResolver;
 use maps::Pid;
 use util::uname_release;
 
@@ -148,85 +145,6 @@ where
     fn addr_file_off(&self, addr: Addr) -> Option<u64>;
     /// Get the file name of the shared object.
     fn get_obj_file_name(&self) -> &Path;
-}
-
-struct KernelResolver {
-    ksymresolver: Option<Rc<KSymResolver>>,
-    kernelresolver: Option<ElfResolver>,
-    kallsyms: PathBuf,
-    kernel_image: PathBuf,
-}
-
-impl KernelResolver {
-    fn new(
-        kallsyms: &Path,
-        kernel_image: &Path,
-        cache_holder: &CacheHolder,
-    ) -> Result<KernelResolver> {
-        let ksymresolver = cache_holder.get_ksym_cache().get_resolver(kallsyms);
-        let kernelresolver = ElfResolver::new(kernel_image, 0, cache_holder);
-
-        if ksymresolver.is_err() && kernelresolver.is_err() {
-            return Err(Error::new(
-                ErrorKind::NotFound,
-                format!(
-                    "can not load {} and {}",
-                    kallsyms.display(),
-                    kernel_image.display()
-                ),
-            ))
-        }
-
-        Ok(KernelResolver {
-            ksymresolver: ksymresolver.ok(),
-            kernelresolver: kernelresolver.ok(),
-            kallsyms: kallsyms.to_path_buf(),
-            kernel_image: kernel_image.to_path_buf(),
-        })
-    }
-}
-
-impl SymResolver for KernelResolver {
-    fn get_address_range(&self) -> (Addr, Addr) {
-        (0xffffffff80000000, 0xffffffffffffffff)
-    }
-
-    fn find_symbols(&self, addr: Addr) -> Vec<(&str, Addr)> {
-        if self.ksymresolver.is_some() {
-            self.ksymresolver.as_ref().unwrap().find_symbols(addr)
-        } else {
-            self.kernelresolver.as_ref().unwrap().find_symbols(addr)
-        }
-    }
-    fn find_address(&self, _name: &str, _opts: &FindAddrOpts) -> Option<Vec<SymbolInfo>> {
-        None
-    }
-    fn find_address_regex(&self, _name: &str, _opts: &FindAddrOpts) -> Option<Vec<SymbolInfo>> {
-        None
-    }
-    fn find_line_info(&self, addr: Addr) -> Option<AddressLineInfo> {
-        self.kernelresolver.as_ref()?;
-        self.kernelresolver.as_ref().unwrap().find_line_info(addr)
-    }
-
-    fn addr_file_off(&self, _addr: Addr) -> Option<u64> {
-        None
-    }
-
-    fn get_obj_file_name(&self) -> &Path {
-        &self.kernel_image
-    }
-}
-
-impl Debug for KernelResolver {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        write!(
-            f,
-            "KernelResolver {} {}",
-            self.kallsyms.display(),
-            self.kernel_image.display()
-        )
-    }
 }
